+++ src/main/java/com/epam/reportportal/cucumber/ScenarioReporter.java
@@ -48,13 +48,17 @@ import java.util.Calendar;
  * @author Sergey_Gvozdyukevich
  */
 public class ScenarioReporter extends AbstractReporter {
+	private static final String RP_STORY_TYPE = "SUITE";
+	private static final String RP_TEST_TYPE = "STORY";
+	private static final String RP_STEP_TYPE = "STEP";
+
 	private static final Logger LOGGER = LoggerFactory.getLogger(ScenarioReporter.class);
 
 	protected Supplier<Maybe<String>> rootSuiteId = Suppliers.memoize(() -> {
 		StartTestItemRQ rq = new StartTestItemRQ();
 		rq.setName("Root User Story");
 		rq.setStartTime(Calendar.getInstance().getTime());
-		rq.setType("STORY");
+		rq.setType(RP_STORY_TYPE);
 		return launch.get().startTestItem(rq);
 	});
 
@@ -105,12 +109,12 @@ public class ScenarioReporter extends AbstractReporter {
 
 	@Override
 	protected String getFeatureTestItemType() {
-		return "TEST";
+		return RP_TEST_TYPE;
 	}
 
 	@Override
 	protected String getScenarioTestItemType() {
-		return "STEP";
+		return RP_STEP_TYPE;
 	}
 
 	@Override