+++ src/main/java/com/epam/reportportal/cucumber/StepReporter.java
@@ -43,6 +43,8 @@ import java.util.Calendar;
  * @author Sergey_Gvozdyukevich
  */
 public class StepReporter extends AbstractReporter {
+	private static final String RP_STORY_TYPE = "STORY";
+	private static final String RP_TEST_TYPE = "SCENARIO";
 
 	public StepReporter() {
 		super();
@@ -98,11 +100,11 @@ public class StepReporter extends AbstractReporter {
 
 	@Override
 	protected String getFeatureTestItemType() {
-		return "SUITE";
+		return RP_STORY_TYPE;
 	}
 
 	@Override
 	protected String getScenarioTestItemType() {
-		return "SCENARIO";
+		return RP_TEST_TYPE;
 	}
 }